Remove ISS Scheduling to Simplify
Most users have a filter for NOAA/M2 and the ISS captures are a bit of a
distraction from the earth-focused captures for the time being. These
can obviously be added back in later, but for the sake of limiting scope
in the refactor of scripts, these are being removed for the time being.