Add Thermal Image Processing for NOAA
Add the ability to create a thermal image for NOAA, and error handling
to ensure any variant to be displayed to the webpanel has an actual
image file to be displayed (e.g. thermal images are only going to occur
going forward - so we don't want an image placeholder to be in place for
older captures that don't have a thermal variant). This logic was built
more generically to strip out any filter that doesn't have an actual
file available.